requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
pymysql==1.1.0  # 纯Python MySQL驱动，无需编译
python-dotenv==1.0.0
sqlalchemy
cryptography